[DROOLS-2911] Adapting to kie-wb-common-command-api refactoring. Still broke.
