+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/resources/org/drools/workbench/screens/scenariosimulation/DroolsWorkbenchScenarioSimulationEditorClient.gwt.xml
@@ -44,5 +44,7 @@
 
   <inherits name="org.guvnor.common.services.project.GuvnorProjectClient" />
 
+  <inherits name="org.kie.workbench.common.command.CommonCommandAPI" />
+
 
 </module>
